Network.AWS.GameLift.CreateBuild

Description

Creates a new Amazon GameLift build from a set of game server binary files stored in an Amazon Simple Storage Service (Amazon S3) location. To use this API call, create a .zip file containing all of the files for the build and store it in an Amazon S3 bucket under your AWS account. For help on packaging your build files and creating a build, see Uploading Your Game to Amazon GameLift .

Important: Use this API action ONLY if you are storing your game build files in an Amazon S3 bucket. To create a build using files stored locally, use the CLI command @upload-build@ , which uploads the build files from a file location you specify.

To create a new build using CreateBuild , identify the storage location and operating system of your game build. You also have the option of specifying a build name and version. If successful, this action creates a new build record with an unique build ID and in INITIALIZED status. Use the API call DescribeBuild to check the status of your build. A build must be in READY status before it can be used to create fleets to host your game.

Build-related operations include:

CreateBuild

ListBuilds

DescribeBuild

UpdateBuild

DeleteBuild

Creating a Request

createBuild :: CreateBuild

Creates a value of CreateBuild with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbStorageLocation - Amazon S3 location of the game build files to be uploaded. The S3 bucket must be owned by the same AWS account that you're using to manage Amazon GameLift. It also must in the same region that you want to create a new build in. Before calling CreateBuild with this location, you must allow Amazon GameLift to access your Amazon S3 bucket (see Create a Build with Files in Amazon S3 ).

cbOperatingSystem - Operating system that the game server binaries are built to run on. This value determines the type of fleet resources that you can use for this build. If your game build contains multiple executables, they all must run on the same operating system.

cbName - Descriptive label that is associated with a build. Build names do not need to be unique. You can use UpdateBuild to change this value later.

cbVersion - Version that is associated with this build. Version strings do not need to be unique. You can use UpdateBuild to change this value later.

data CreateBuild

Represents the input for a request action.

See: createBuild smart constructor.

Request Lenses

cbStorageLocation :: Lens' CreateBuild (Maybe S3Location)

Amazon S3 location of the game build files to be uploaded. The S3 bucket must be owned by the same AWS account that you're using to manage Amazon GameLift. It also must in the same region that you want to create a new build in. Before calling CreateBuild with this location, you must allow Amazon GameLift to access your Amazon S3 bucket (see Create a Build with Files in Amazon S3 ).

cbOperatingSystem :: Lens' CreateBuild (Maybe OperatingSystem)

Operating system that the game server binaries are built to run on. This value determines the type of fleet resources that you can use for this build. If your game build contains multiple executables, they all must run on the same operating system.

cbName :: Lens' CreateBuild (Maybe Text)

Descriptive label that is associated with a build. Build names do not need to be unique. You can use UpdateBuild to change this value later.

cbVersion :: Lens' CreateBuild (Maybe Text)

Version that is associated with this build. Version strings do not need to be unique. You can use UpdateBuild to change this value later.

Destructuring the Response

createBuildResponse

Arguments

:: Int
cbrsResponseStatus

-> CreateBuildResponse

Creates a value of CreateBuildResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbrsStorageLocation - Amazon S3 location specified in the request.

cbrsUploadCredentials - This element is not currently in use.

cbrsBuild - The newly created build record, including a unique build ID and status.

cbrsResponseStatus - -- | The response status code.

data CreateBuildResponse

Represents the returned data in response to a request action.

See: createBuildResponse smart constructor.

Response Lenses

cbrsStorageLocation :: Lens' CreateBuildResponse (Maybe S3Location)

Amazon S3 location specified in the request.

cbrsUploadCredentials :: Lens' CreateBuildResponse (Maybe AWSCredentials)

This element is not currently in use.

cbrsBuild :: Lens' CreateBuildResponse (Maybe Build)

The newly created build record, including a unique build ID and status.

cbrsResponseStatus :: Lens' CreateBuildResponse Int

- | The response status code.