Network.AWS.GameLift.CreateGameSession

Description

Creates a multiplayer game session for players. This action creates a game session record and assigns an available server process in the specified fleet to host the game session. A fleet must have an ACTIVE status before a game session can be created in it.

To create a game session, specify either fleet ID or alias ID and indicate a maximum number of players to allow in the game session. You can also provide a name and game-specific properties for this game session. If successful, a GameSession object is returned containing the game session properties and other settings you specified.

Idempotency tokens. You can add a token that uniquely identifies game session requests. This is useful for ensuring that game session requests are idempotent. Multiple requests with the same idempotency token are processed only once; subsequent requests return the original result. All response values are the same with the exception of game session status, which may change.

Resource creation limits. If you are creating a game session on a fleet with a resource creation limit policy in force, then you must specify a creator ID. Without this ID, Amazon GameLift has no way to evaluate the policy for this new game session request.

Player acceptance policy. By default, newly created game sessions are open to new players. You can restrict new player access by using UpdateGameSession to change the game session's player session creation policy.

Game session logs. Logs are retained for all active game sessions for 14 days. To access the logs, call GetGameSessionLogUrl to download the log files.

Available in Amazon GameLift Local.

Game-session-related operations include:

CreateGameSession

DescribeGameSessions

DescribeGameSessionDetails

SearchGameSessions

UpdateGameSession

GetGameSessionLogUrl

Game session placements

StartGameSessionPlacement

DescribeGameSessionPlacement

StopGameSessionPlacement

Creating a Request

createGameSession

Arguments

:: Natural
cgsMaximumPlayerSessionCount

-> CreateGameSession

Creates a value of CreateGameSession with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgsIdempotencyToken - Custom string that uniquely identifies a request for a new game session. Maximum token length is 48 characters. If provided, this string is included in the new game session's ID. (A game session ID has the following format: arn:aws:gamelift:region::gamesessionIDID string or idempotency token .) Idempotency tokens remain in use for 30 days after a game session has ended; game session objects are retained for this time period and then deleted.

cgsGameProperties - Set of developer-defined properties for a game session, formatted as a set of type:value pairs. These properties are included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ).

cgsGameSessionId - This parameter is no longer preferred. Please use IdempotencyToken instead. Custom string that uniquely identifies a request for a new game session. Maximum token length is 48 characters. If provided, this string is included in the new game session's ID. (A game session ID has the following format: arn:aws:gamelift:region::gamesessionIDID string or idempotency token .)

cgsAliasId - Unique identifier for an alias associated with the fleet to create a game session in. Each request must reference either a fleet ID or alias ID, but not both.

cgsName - Descriptive label that is associated with a game session. Session names do not need to be unique.

cgsGameSessionData - Set of developer-defined game session properties, formatted as a single string value. This data is included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ).

cgsFleetId - Unique identifier for a fleet to create a game session in. Each request must reference either a fleet ID or alias ID, but not both.

cgsCreatorId - Unique identifier for a player or entity creating the game session. This ID is used to enforce a resource protection policy (if one exists) that limits the number of concurrent active game sessions one player can have.

cgsMaximumPlayerSessionCount - Maximum number of players that can be connected simultaneously to the game session.

data CreateGameSession

Represents the input for a request action.

See: createGameSession smart constructor.

Request Lenses

cgsIdempotencyToken :: Lens' CreateGameSession (Maybe Text)

Custom string that uniquely identifies a request for a new game session. Maximum token length is 48 characters. If provided, this string is included in the new game session's ID. (A game session ID has the following format: arn:aws:gamelift:region::gamesessionIDID string or idempotency token .) Idempotency tokens remain in use for 30 days after a game session has ended; game session objects are retained for this time period and then deleted.

cgsGameProperties :: Lens' CreateGameSession [GameProperty]

Set of developer-defined properties for a game session, formatted as a set of type:value pairs. These properties are included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ).

cgsGameSessionId :: Lens' CreateGameSession (Maybe Text)

This parameter is no longer preferred. Please use IdempotencyToken instead. Custom string that uniquely identifies a request for a new game session. Maximum token length is 48 characters. If provided, this string is included in the new game session's ID. (A game session ID has the following format: arn:aws:gamelift:region::gamesessionIDID string or idempotency token .)

cgsAliasId :: Lens' CreateGameSession (Maybe Text)

Unique identifier for an alias associated with the fleet to create a game session in. Each request must reference either a fleet ID or alias ID, but not both.

cgsName :: Lens' CreateGameSession (Maybe Text)

Descriptive label that is associated with a game session. Session names do not need to be unique.

cgsGameSessionData :: Lens' CreateGameSession (Maybe Text)

Set of developer-defined game session properties, formatted as a single string value. This data is included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ).

cgsFleetId :: Lens' CreateGameSession (Maybe Text)

Unique identifier for a fleet to create a game session in. Each request must reference either a fleet ID or alias ID, but not both.

cgsCreatorId :: Lens' CreateGameSession (Maybe Text)

Unique identifier for a player or entity creating the game session. This ID is used to enforce a resource protection policy (if one exists) that limits the number of concurrent active game sessions one player can have.

cgsMaximumPlayerSessionCount :: Lens' CreateGameSession Natural

Maximum number of players that can be connected simultaneously to the game session.

Destructuring the Response

createGameSessionResponse

Arguments

:: Int
cgsrsResponseStatus

-> CreateGameSessionResponse

Creates a value of CreateGameSessionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgsrsGameSession - Object that describes the newly created game session record.

cgsrsResponseStatus - -- | The response status code.

data CreateGameSessionResponse

Represents the returned data in response to a request action.

See: createGameSessionResponse smart constructor.

Response Lenses

cgsrsGameSession :: Lens' CreateGameSessionResponse (Maybe GameSession)

Object that describes the newly created game session record.

cgsrsResponseStatus :: Lens' CreateGameSessionResponse Int

- | The response status code.