Network.AWS.GameLift.CreateGameSessionQueue

Description

Establishes a new queue for processing requests to place new game sessions. A queue identifies where new game sessions can be hosted -- by specifying a list of destinations (fleets or aliases) -- and how long requests can wait in the queue before timing out. You can set up a queue to try to place game sessions on fleets in multiple regions. To add placement requests to a queue, call StartGameSessionPlacement and reference the queue name.

Destination order. When processing a request for a game session, Amazon GameLift tries each destination in order until it finds one with available resources to host the new game session. A queue's default order is determined by how destinations are listed. The default order is overridden when a game session placement request provides player latency information. Player latency information enables Amazon GameLift to prioritize destinations where players report the lowest average latency, as a result placing the new game session where the majority of players will have the best possible gameplay experience.

Player latency policies. For placement requests containing player latency information, use player latency policies to protect individual players from very high latencies. With a latency cap, even when a destination can deliver a low latency for most players, the game is not placed where any individual player is reporting latency higher than a policy's maximum. A queue can have multiple latency policies, which are enforced consecutively starting with the policy with the lowest latency cap. Use multiple policies to gradually relax latency controls; for example, you might set a policy with a low latency cap for the first 60 seconds, a second policy with a higher cap for the next 60 seconds, etc.

To create a new queue, provide a name, timeout value, a list of destinations and, if desired, a set of latency policies. If successful, a new queue object is returned.

Queue-related operations include:

CreateGameSessionQueue

DescribeGameSessionQueues

UpdateGameSessionQueue

DeleteGameSessionQueue

Creating a Request

createGameSessionQueue

Arguments

:: Text
cgsqName

-> CreateGameSessionQueue

Creates a value of CreateGameSessionQueue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgsqPlayerLatencyPolicies - Collection of latency policies to apply when processing game sessions placement requests with player latency information. Multiple policies are evaluated in order of the maximum latency value, starting with the lowest latency values. With just one policy, it is enforced at the start of the game session placement for the duration period. With multiple policies, each policy is enforced consecutively for its duration period. For example, a queue might enforce a 60-second policy followed by a 120-second policy, and then no policy for the remainder of the placement. A player latency policy must set a value for MaximumIndividualPlayerLatencyMilliseconds; if none is set, this API requests will fail.

cgsqTimeoutInSeconds - Maximum time, in seconds, that a new game session placement request remains in the queue. When a request exceeds this time, the game session placement changes to a TIMED_OUT status.

cgsqDestinations - List of fleets that can be used to fulfill game session placement requests in the queue. Fleets are identified by either a fleet ARN or a fleet alias ARN. Destinations are listed in default preference order.

cgsqName - Descriptive label that is associated with game session queue. Queue names must be unique within each region.

data CreateGameSessionQueue

Represents the input for a request action.

See: createGameSessionQueue smart constructor.

Request Lenses

cgsqPlayerLatencyPolicies :: Lens' CreateGameSessionQueue [PlayerLatencyPolicy]

Collection of latency policies to apply when processing game sessions placement requests with player latency information. Multiple policies are evaluated in order of the maximum latency value, starting with the lowest latency values. With just one policy, it is enforced at the start of the game session placement for the duration period. With multiple policies, each policy is enforced consecutively for its duration period. For example, a queue might enforce a 60-second policy followed by a 120-second policy, and then no policy for the remainder of the placement. A player latency policy must set a value for MaximumIndividualPlayerLatencyMilliseconds; if none is set, this API requests will fail.

cgsqTimeoutInSeconds :: Lens' CreateGameSessionQueue (Maybe Natural)

Maximum time, in seconds, that a new game session placement request remains in the queue. When a request exceeds this time, the game session placement changes to a TIMED_OUT status.

cgsqDestinations :: Lens' CreateGameSessionQueue [GameSessionQueueDestination]

List of fleets that can be used to fulfill game session placement requests in the queue. Fleets are identified by either a fleet ARN or a fleet alias ARN. Destinations are listed in default preference order.

cgsqName :: Lens' CreateGameSessionQueue Text

Descriptive label that is associated with game session queue. Queue names must be unique within each region.

Destructuring the Response

createGameSessionQueueResponse

Arguments

:: Int
cgsqrsResponseStatus

-> CreateGameSessionQueueResponse

Creates a value of CreateGameSessionQueueResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgsqrsGameSessionQueue - Object that describes the newly created game session queue.

cgsqrsResponseStatus - -- | The response status code.

data CreateGameSessionQueueResponse

Represents the returned data in response to a request action.

See: createGameSessionQueueResponse smart constructor.

Response Lenses

cgsqrsGameSessionQueue :: Lens' CreateGameSessionQueueResponse (Maybe GameSessionQueue)

Object that describes the newly created game session queue.

cgsqrsResponseStatus :: Lens' CreateGameSessionQueueResponse Int

- | The response status code.