Network.AWS.GameLift.CreateMatchmakingConfiguration

Description

Defines a new matchmaking configuration for use with FlexMatch. A matchmaking configuration sets out guidelines for matching players and getting the matches into games. You can set up multiple matchmaking configurations to handle the scenarios needed for your game. Each matchmaking request (StartMatchmaking ) specifies a configuration for the match and provides player attributes to support the configuration being used.

To create a matchmaking configuration, at a minimum you must specify the following: configuration name; a rule set that governs how to evaluate players and find acceptable matches; a game session queue to use when placing a new game session for the match; and the maximum time allowed for a matchmaking attempt.

Player acceptance -- In each configuration, you have the option to require that all players accept participation in a proposed match. To enable this feature, set AcceptanceRequired to true and specify a time limit for player acceptance. Players have the option to accept or reject a proposed match, and a match does not move ahead to game session placement unless all matched players accept.

Matchmaking status notification -- There are two ways to track the progress of matchmaking tickets: (1) polling ticket status with DescribeMatchmaking ; or (2) receiving notifications with Amazon Simple Notification Service (SNS). To use notifications, you first need to set up an SNS topic to receive the notifications, and provide the topic ARN in the matchmaking configuration (see Setting up Notifications for Matchmaking ). Since notifications promise only "best effort" delivery, we recommend calling DescribeMatchmaking if no notifications are received within 30 seconds.

Operations related to match configurations and rule sets include:

CreateMatchmakingConfiguration

DescribeMatchmakingConfigurations

UpdateMatchmakingConfiguration

DeleteMatchmakingConfiguration

CreateMatchmakingRuleSet

DescribeMatchmakingRuleSets

ValidateMatchmakingRuleSet

Creating a Request

createMatchmakingConfiguration

Arguments

:: Text
cmcName

-> Natural
cmcRequestTimeoutSeconds

-> Bool
cmcAcceptanceRequired

-> Text
cmcRuleSetName

-> CreateMatchmakingConfiguration

Creates a value of CreateMatchmakingConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmcGameProperties - Set of developer-defined properties for a game session, formatted as a set of type:value pairs. These properties are included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ). This information is added to the new GameSession object that is created for a successful match.

cmcAcceptanceTimeoutSeconds - Length of time (in seconds) to wait for players to accept a proposed match. If any player rejects the match or fails to accept before the timeout, the ticket continues to look for an acceptable match.

cmcNotificationTarget - SNS topic ARN that is set up to receive matchmaking notifications.

cmcCustomEventData - Information to attached to all events related to the matchmaking configuration.

cmcGameSessionData - Set of developer-defined game session properties, formatted as a single string value. This data is included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ). This information is added to the new GameSession object that is created for a successful match.

cmcDescription - Meaningful description of the matchmaking configuration.

cmcAdditionalPlayerCount - Number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match.

cmcName - Unique identifier for a matchmaking configuration. This name is used to identify the configuration associated with a matchmaking request or ticket.

cmcGameSessionQueueARNs - Amazon Resource Name (ARN ) that is assigned to a game session queue and uniquely identifies it. Format is arn:aws:gamelift:region::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912 . These queues are used when placing game sessions for matches that are created with this matchmaking configuration. Queues can be located in any region.

cmcRequestTimeoutSeconds - Maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that time out can be resubmitted as needed.

cmcAcceptanceRequired - Flag that determines whether or not a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE.

cmcRuleSetName - Unique identifier for a matchmaking rule set to use with this configuration. A matchmaking configuration can only use rule sets that are defined in the same region.

data CreateMatchmakingConfiguration

Represents the input for a request action.

See: createMatchmakingConfiguration smart constructor.

Request Lenses

cmcGameProperties :: Lens' CreateMatchmakingConfiguration [GameProperty]

Set of developer-defined properties for a game session, formatted as a set of type:value pairs. These properties are included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ). This information is added to the new GameSession object that is created for a successful match.

cmcAcceptanceTimeoutSeconds :: Lens' CreateMatchmakingConfiguration (Maybe Natural)

Length of time (in seconds) to wait for players to accept a proposed match. If any player rejects the match or fails to accept before the timeout, the ticket continues to look for an acceptable match.

cmcNotificationTarget :: Lens' CreateMatchmakingConfiguration (Maybe Text)

SNS topic ARN that is set up to receive matchmaking notifications.

cmcCustomEventData :: Lens' CreateMatchmakingConfiguration (Maybe Text)

Information to attached to all events related to the matchmaking configuration.

cmcGameSessionData :: Lens' CreateMatchmakingConfiguration (Maybe Text)

Set of developer-defined game session properties, formatted as a single string value. This data is included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ). This information is added to the new GameSession object that is created for a successful match.

cmcDescription :: Lens' CreateMatchmakingConfiguration (Maybe Text)

Meaningful description of the matchmaking configuration.

cmcAdditionalPlayerCount :: Lens' CreateMatchmakingConfiguration (Maybe Natural)

Number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match.

cmcName :: Lens' CreateMatchmakingConfiguration Text

Unique identifier for a matchmaking configuration. This name is used to identify the configuration associated with a matchmaking request or ticket.

cmcGameSessionQueueARNs :: Lens' CreateMatchmakingConfiguration [Text]

Amazon Resource Name (ARN ) that is assigned to a game session queue and uniquely identifies it. Format is arn:aws:gamelift:region::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912 . These queues are used when placing game sessions for matches that are created with this matchmaking configuration. Queues can be located in any region.

cmcRequestTimeoutSeconds :: Lens' CreateMatchmakingConfiguration Natural

Maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that time out can be resubmitted as needed.

cmcAcceptanceRequired :: Lens' CreateMatchmakingConfiguration Bool

Flag that determines whether or not a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE.

cmcRuleSetName :: Lens' CreateMatchmakingConfiguration Text

Unique identifier for a matchmaking rule set to use with this configuration. A matchmaking configuration can only use rule sets that are defined in the same region.

Destructuring the Response

createMatchmakingConfigurationResponse

Arguments

:: Int
cmcrsResponseStatus

-> CreateMatchmakingConfigurationResponse

Creates a value of CreateMatchmakingConfigurationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmcrsConfiguration - Object that describes the newly created matchmaking configuration.

cmcrsResponseStatus - -- | The response status code.

data CreateMatchmakingConfigurationResponse

Represents the returned data in response to a request action.

See: createMatchmakingConfigurationResponse smart constructor.

Response Lenses

cmcrsConfiguration :: Lens' CreateMatchmakingConfigurationResponse (Maybe MatchmakingConfiguration)

Object that describes the newly created matchmaking configuration.

cmcrsResponseStatus :: Lens' CreateMatchmakingConfigurationResponse Int

- | The response status code.