Network.AWS.GameLift.CreateMatchmakingRuleSet

Description

Creates a new rule set for FlexMatch matchmaking. A rule set describes the type of match to create, such as the number and size of teams, and sets the parameters for acceptable player matches, such as minimum skill level or character type. Rule sets are used in matchmaking configurations, which define how matchmaking requests are handled. Each MatchmakingConfiguration uses one rule set; you can set up multiple rule sets to handle the scenarios that suit your game (such as for different game modes), and create a separate matchmaking configuration for each rule set. See additional information on rule set content in the MatchmakingRuleSet structure. For help creating rule sets, including useful examples, see the topic Adding FlexMatch to Your Game .

Once created, matchmaking rule sets cannot be changed or deleted, so we recommend checking the rule set syntax using ValidateMatchmakingRuleSet before creating the rule set.

To create a matchmaking rule set, provide the set of rules and a unique name. Rule sets must be defined in the same region as the matchmaking configuration they will be used with. Rule sets cannot be edited or deleted. If you need to change a rule set, create a new one with the necessary edits and then update matchmaking configurations to use the new rule set.

Operations related to match configurations and rule sets include:

CreateMatchmakingConfiguration

DescribeMatchmakingConfigurations

UpdateMatchmakingConfiguration

DeleteMatchmakingConfiguration

CreateMatchmakingRuleSet

DescribeMatchmakingRuleSets

ValidateMatchmakingRuleSet

Creating a Request

createMatchmakingRuleSet

Arguments

:: Text
cmrsName

-> Text
cmrsRuleSetBody

-> CreateMatchmakingRuleSet

Creates a value of CreateMatchmakingRuleSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmrsName - Unique identifier for a matchmaking rule set. This name is used to identify the rule set associated with a matchmaking configuration.

cmrsRuleSetBody - Collection of matchmaking rules, formatted as a JSON string. (Note that comments are not allowed in JSON, but most elements support a description field.)

data CreateMatchmakingRuleSet

Represents the input for a request action.

See: createMatchmakingRuleSet smart constructor.

Request Lenses

cmrsName :: Lens' CreateMatchmakingRuleSet Text

Unique identifier for a matchmaking rule set. This name is used to identify the rule set associated with a matchmaking configuration.

cmrsRuleSetBody :: Lens' CreateMatchmakingRuleSet Text

Collection of matchmaking rules, formatted as a JSON string. (Note that comments are not allowed in JSON, but most elements support a description field.)

Destructuring the Response

createMatchmakingRuleSetResponse

Arguments

:: Int
cmrsrsResponseStatus

-> MatchmakingRuleSet
cmrsrsRuleSet

-> CreateMatchmakingRuleSetResponse

Creates a value of CreateMatchmakingRuleSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmrsrsResponseStatus - -- | The response status code.

cmrsrsRuleSet - Object that describes the newly created matchmaking rule set.

data CreateMatchmakingRuleSetResponse

Represents the returned data in response to a request action.

See: createMatchmakingRuleSetResponse smart constructor.

Response Lenses

cmrsrsResponseStatus :: Lens' CreateMatchmakingRuleSetResponse Int

- | The response status code.

cmrsrsRuleSet :: Lens' CreateMatchmakingRuleSetResponse MatchmakingRuleSet

Object that describes the newly created matchmaking rule set.