Network.AWS.GameLift.DeleteGameSessionQueue

Description

Deletes a game session queue. This action means that any StartGameSessionPlacement requests that reference this queue will fail. To delete a queue, specify the queue name.

Queue-related operations include:

CreateGameSessionQueue

DescribeGameSessionQueues

UpdateGameSessionQueue

DeleteGameSessionQueue

Creating a Request

deleteGameSessionQueue

Arguments

:: Text
dgsqName

-> DeleteGameSessionQueue

Creates a value of DeleteGameSessionQueue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dgsqName - Descriptive label that is associated with game session queue. Queue names must be unique within each region.

data DeleteGameSessionQueue

Represents the input for a request action.

See: deleteGameSessionQueue smart constructor.

Request Lenses

dgsqName :: Lens' DeleteGameSessionQueue Text

Descriptive label that is associated with game session queue. Queue names must be unique within each region.

Destructuring the Response

deleteGameSessionQueueResponse

Arguments

:: Int
dgsqrsResponseStatus

-> DeleteGameSessionQueueResponse

Creates a value of DeleteGameSessionQueueResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dgsqrsResponseStatus - -- | The response status code.

data DeleteGameSessionQueueResponse

See: deleteGameSessionQueueResponse smart constructor.

Response Lenses

dgsqrsResponseStatus :: Lens' DeleteGameSessionQueueResponse Int

- | The response status code.