Network.AWS.GameLift.DeleteMatchmakingConfiguration

Description

Permanently removes a FlexMatch matchmaking configuration. To delete, specify the configuration name. A matchmaking configuration cannot be deleted if it is being used in any active matchmaking tickets.

Operations related to match configurations and rule sets include:

CreateMatchmakingConfiguration

DescribeMatchmakingConfigurations

UpdateMatchmakingConfiguration

DeleteMatchmakingConfiguration

CreateMatchmakingRuleSet

DescribeMatchmakingRuleSets

ValidateMatchmakingRuleSet

Creating a Request

deleteMatchmakingConfiguration

Arguments

:: Text
dmcName

-> DeleteMatchmakingConfiguration

Creates a value of DeleteMatchmakingConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmcName - Unique identifier for a matchmaking configuration

data DeleteMatchmakingConfiguration

Represents the input for a request action.

See: deleteMatchmakingConfiguration smart constructor.

Request Lenses

dmcName :: Lens' DeleteMatchmakingConfiguration Text

Unique identifier for a matchmaking configuration

Destructuring the Response

deleteMatchmakingConfigurationResponse

Arguments

:: Int
dmcrsResponseStatus

-> DeleteMatchmakingConfigurationResponse

Creates a value of DeleteMatchmakingConfigurationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmcrsResponseStatus - -- | The response status code.

data DeleteMatchmakingConfigurationResponse

See: deleteMatchmakingConfigurationResponse smart constructor.

Response Lenses

dmcrsResponseStatus :: Lens' DeleteMatchmakingConfigurationResponse Int

- | The response status code.