Network.AWS.GameLift.DescribeAlias

Description

Retrieves properties for an alias. This operation returns all alias metadata and settings. To get an alias's target fleet ID only, use ResolveAlias .

To get alias properties, specify the alias ID. If successful, the requested alias record is returned.

Alias-related operations include:

CreateAlias

ListAliases

DescribeAlias

UpdateAlias

DeleteAlias

ResolveAlias

Creating a Request

describeAlias

Arguments

:: Text
dAliasId

-> DescribeAlias

Creates a value of DescribeAlias with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dAliasId - Unique identifier for a fleet alias. Specify the alias you want to retrieve.

data DescribeAlias

Represents the input for a request action.

See: describeAlias smart constructor.

Request Lenses

dAliasId :: Lens' DescribeAlias Text

Unique identifier for a fleet alias. Specify the alias you want to retrieve.

Destructuring the Response

describeAliasResponse

Arguments

:: Int
darsResponseStatus

-> DescribeAliasResponse

Creates a value of DescribeAliasResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

darsAlias - Object that contains the requested alias.

darsResponseStatus - -- | The response status code.

data DescribeAliasResponse

Represents the returned data in response to a request action.

See: describeAliasResponse smart constructor.

Response Lenses

darsAlias :: Lens' DescribeAliasResponse (Maybe Alias)

Object that contains the requested alias.

darsResponseStatus :: Lens' DescribeAliasResponse Int

- | The response status code.