Network.AWS.GameLift.DescribeFleetAttributes

Description

Retrieves fleet properties, including metadata, status, and configuration, for one or more fleets. You can request attributes for all fleets, or specify a list of one or more fleet IDs. When requesting multiple fleets, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a FleetAttributes object is returned for each requested fleet ID. When specifying a list of fleet IDs, attribute objects are returned only for fleets that currently exist.

Fleet-related operations include:

CreateFleet

ListFleets

Describe fleets:

DescribeFleetAttributes

DescribeFleetPortSettings

DescribeFleetUtilization

DescribeRuntimeConfiguration

DescribeFleetEvents

Update fleets:

UpdateFleetAttributes

UpdateFleetCapacity

UpdateFleetPortSettings

UpdateRuntimeConfiguration

Manage fleet capacity:

DescribeFleetCapacity

UpdateFleetCapacity

PutScalingPolicy (automatic scaling)

DescribeScalingPolicies (automatic scaling)

DeleteScalingPolicy (automatic scaling)

DescribeEC2InstanceLimits

DeleteFleet

Creating a Request

describeFleetAttributes :: DescribeFleetAttributes

Creates a value of DescribeFleetAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfaNextToken - Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value. This parameter is ignored when the request specifies one or a list of fleet IDs.

dfaLimit - Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is ignored when the request specifies one or a list of fleet IDs.

dfaFleetIds - Unique identifier for a fleet(s) to retrieve attributes for. To request attributes for all fleets, leave this parameter empty.

data DescribeFleetAttributes

Represents the input for a request action.

See: describeFleetAttributes smart constructor.

Request Lenses

dfaNextToken :: Lens' DescribeFleetAttributes (Maybe Text)

Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value. This parameter is ignored when the request specifies one or a list of fleet IDs.

dfaLimit :: Lens' DescribeFleetAttributes (Maybe Natural)

Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is ignored when the request specifies one or a list of fleet IDs.

dfaFleetIds :: Lens' DescribeFleetAttributes (Maybe (NonEmpty Text))

Unique identifier for a fleet(s) to retrieve attributes for. To request attributes for all fleets, leave this parameter empty.

Destructuring the Response

describeFleetAttributesResponse

Arguments

:: Int
dfarsResponseStatus

-> DescribeFleetAttributesResponse

Creates a value of DescribeFleetAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfarsNextToken - Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dfarsFleetAttributes - Collection of objects containing attribute metadata for each requested fleet ID.

dfarsResponseStatus - -- | The response status code.

data DescribeFleetAttributesResponse

Represents the returned data in response to a request action.

See: describeFleetAttributesResponse smart constructor.

Response Lenses

dfarsNextToken :: Lens' DescribeFleetAttributesResponse (Maybe Text)

Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dfarsFleetAttributes :: Lens' DescribeFleetAttributesResponse [FleetAttributes]

Collection of objects containing attribute metadata for each requested fleet ID.

dfarsResponseStatus :: Lens' DescribeFleetAttributesResponse Int

- | The response status code.