Network.AWS.GameLift.DescribeFleetEvents

Description

Retrieves entries from the specified fleet's event log. You can specify a time range to limit the result set. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, a collection of event log entries matching the request are returned.

Fleet-related operations include:

CreateFleet

ListFleets

Describe fleets:

DescribeFleetAttributes

DescribeFleetPortSettings

DescribeFleetUtilization

DescribeRuntimeConfiguration

DescribeFleetEvents

Update fleets:

UpdateFleetAttributes

UpdateFleetCapacity

UpdateFleetPortSettings

UpdateRuntimeConfiguration

Manage fleet capacity:

DescribeFleetCapacity

UpdateFleetCapacity

PutScalingPolicy (automatic scaling)

DescribeScalingPolicies (automatic scaling)

DeleteScalingPolicy (automatic scaling)

DescribeEC2InstanceLimits

DeleteFleet

Creating a Request

describeFleetEvents

Arguments

:: Text
dfeFleetId

-> DescribeFleetEvents

Creates a value of DescribeFleetEvents with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfeStartTime - Earliest date to retrieve event logs for. If no start time is specified, this call returns entries starting from when the fleet was created to the specified end time. Format is a number expressed in Unix time as milliseconds (ex: "1469498468.057").

dfeNextToken - Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

dfeEndTime - Most recent date to retrieve event logs for. If no end time is specified, this call returns entries from the specified start time up to the present. Format is a number expressed in Unix time as milliseconds (ex: "1469498468.057").

dfeLimit - Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

dfeFleetId - Unique identifier for a fleet to get event logs for.

data DescribeFleetEvents

Represents the input for a request action.

See: describeFleetEvents smart constructor.

Request Lenses

dfeStartTime :: Lens' DescribeFleetEvents (Maybe UTCTime)

Earliest date to retrieve event logs for. If no start time is specified, this call returns entries starting from when the fleet was created to the specified end time. Format is a number expressed in Unix time as milliseconds (ex: "1469498468.057").

dfeNextToken :: Lens' DescribeFleetEvents (Maybe Text)

Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

dfeEndTime :: Lens' DescribeFleetEvents (Maybe UTCTime)

Most recent date to retrieve event logs for. If no end time is specified, this call returns entries from the specified start time up to the present. Format is a number expressed in Unix time as milliseconds (ex: "1469498468.057").

dfeLimit :: Lens' DescribeFleetEvents (Maybe Natural)

Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

dfeFleetId :: Lens' DescribeFleetEvents Text

Unique identifier for a fleet to get event logs for.

Destructuring the Response

describeFleetEventsResponse

Arguments

:: Int
dfersResponseStatus

-> DescribeFleetEventsResponse

Creates a value of DescribeFleetEventsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfersNextToken - Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dfersEvents - Collection of objects containing event log entries for the specified fleet.

dfersResponseStatus - -- | The response status code.

data DescribeFleetEventsResponse

Represents the returned data in response to a request action.

See: describeFleetEventsResponse smart constructor.

Response Lenses

dfersNextToken :: Lens' DescribeFleetEventsResponse (Maybe Text)

Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dfersEvents :: Lens' DescribeFleetEventsResponse [Event]

Collection of objects containing event log entries for the specified fleet.

dfersResponseStatus :: Lens' DescribeFleetEventsResponse Int

- | The response status code.