Network.AWS.GameLift.DescribeFleetUtilization

Description

Retrieves utilization statistics for one or more fleets. You can request utilization data for all fleets, or specify a list of one or more fleet IDs. When requesting multiple fleets, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a FleetUtilization object is returned for each requested fleet ID. When specifying a list of fleet IDs, utilization objects are returned only for fleets that currently exist.

Fleet-related operations include:

CreateFleet

ListFleets

Describe fleets:

DescribeFleetAttributes

DescribeFleetPortSettings

DescribeFleetUtilization

DescribeRuntimeConfiguration

DescribeFleetEvents

Update fleets:

UpdateFleetAttributes

UpdateFleetCapacity

UpdateFleetPortSettings

UpdateRuntimeConfiguration

Manage fleet capacity:

DescribeFleetCapacity

UpdateFleetCapacity

PutScalingPolicy (automatic scaling)

DescribeScalingPolicies (automatic scaling)

DeleteScalingPolicy (automatic scaling)

DescribeEC2InstanceLimits

DeleteFleet

Creating a Request

describeFleetUtilization :: DescribeFleetUtilization

Creates a value of DescribeFleetUtilization with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfuNextToken - Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value. This parameter is ignored when the request specifies one or a list of fleet IDs.

dfuLimit - Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is ignored when the request specifies one or a list of fleet IDs.

dfuFleetIds - Unique identifier for a fleet(s) to retrieve utilization data for. To request utilization data for all fleets, leave this parameter empty.

data DescribeFleetUtilization

Represents the input for a request action.

See: describeFleetUtilization smart constructor.

Request Lenses

dfuNextToken :: Lens' DescribeFleetUtilization (Maybe Text)

Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value. This parameter is ignored when the request specifies one or a list of fleet IDs.

dfuLimit :: Lens' DescribeFleetUtilization (Maybe Natural)

Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is ignored when the request specifies one or a list of fleet IDs.

dfuFleetIds :: Lens' DescribeFleetUtilization (Maybe (NonEmpty Text))

Unique identifier for a fleet(s) to retrieve utilization data for. To request utilization data for all fleets, leave this parameter empty.

Destructuring the Response

describeFleetUtilizationResponse

Arguments

:: Int
dfursResponseStatus

-> DescribeFleetUtilizationResponse

Creates a value of DescribeFleetUtilizationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfursNextToken - Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dfursFleetUtilization - Collection of objects containing utilization information for each requested fleet ID.

dfursResponseStatus - -- | The response status code.

data DescribeFleetUtilizationResponse

Represents the returned data in response to a request action.

See: describeFleetUtilizationResponse smart constructor.

Response Lenses

dfursNextToken :: Lens' DescribeFleetUtilizationResponse (Maybe Text)

Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dfursFleetUtilization :: Lens' DescribeFleetUtilizationResponse [FleetUtilization]

Collection of objects containing utilization information for each requested fleet ID.

dfursResponseStatus :: Lens' DescribeFleetUtilizationResponse Int

- | The response status code.