Network.AWS.GameLift.DescribeGameSessionDetails

Description

Retrieves properties, including the protection policy in force, for one or more game sessions. This action can be used in several ways: (1) provide a GameSessionId or GameSessionArn to request details for a specific game session; (2) provide either a FleetId or an AliasId to request properties for all game sessions running on a fleet.

To get game session record(s), specify just one of the following: game session ID, fleet ID, or alias ID. You can filter this request by game session status. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, a GameSessionDetail object is returned for each session matching the request.

Game-session-related operations include:

CreateGameSession

DescribeGameSessions

DescribeGameSessionDetails

SearchGameSessions

UpdateGameSession

GetGameSessionLogUrl

Game session placements

StartGameSessionPlacement

DescribeGameSessionPlacement

StopGameSessionPlacement

Creating a Request

describeGameSessionDetails :: DescribeGameSessionDetails

Creates a value of DescribeGameSessionDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dgsdGameSessionId - Unique identifier for the game session to retrieve.

dgsdAliasId - Unique identifier for an alias associated with the fleet to retrieve all game sessions for.

dgsdNextToken - Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

dgsdStatusFilter - Game session status to filter results on. Possible game session statuses include ACTIVE , TERMINATED , ACTIVATING and TERMINATING (the last two are transitory).

dgsdLimit - Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

dgsdFleetId - Unique identifier for a fleet to retrieve all game sessions active on the fleet.

data DescribeGameSessionDetails

Represents the input for a request action.

See: describeGameSessionDetails smart constructor.

Request Lenses

dgsdGameSessionId :: Lens' DescribeGameSessionDetails (Maybe Text)

Unique identifier for the game session to retrieve.

dgsdAliasId :: Lens' DescribeGameSessionDetails (Maybe Text)

Unique identifier for an alias associated with the fleet to retrieve all game sessions for.

dgsdNextToken :: Lens' DescribeGameSessionDetails (Maybe Text)

Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

dgsdStatusFilter :: Lens' DescribeGameSessionDetails (Maybe Text)

Game session status to filter results on. Possible game session statuses include ACTIVE , TERMINATED , ACTIVATING and TERMINATING (the last two are transitory).

dgsdLimit :: Lens' DescribeGameSessionDetails (Maybe Natural)

Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

dgsdFleetId :: Lens' DescribeGameSessionDetails (Maybe Text)

Unique identifier for a fleet to retrieve all game sessions active on the fleet.

Destructuring the Response

describeGameSessionDetailsResponse

Arguments

:: Int
dgsdrsResponseStatus

-> DescribeGameSessionDetailsResponse

Creates a value of DescribeGameSessionDetailsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dgsdrsGameSessionDetails - Collection of objects containing game session properties and the protection policy currently in force for each session matching the request.

dgsdrsNextToken - Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dgsdrsResponseStatus - -- | The response status code.

data DescribeGameSessionDetailsResponse

Represents the returned data in response to a request action.

See: describeGameSessionDetailsResponse smart constructor.

Response Lenses

dgsdrsGameSessionDetails :: Lens' DescribeGameSessionDetailsResponse [GameSessionDetail]

Collection of objects containing game session properties and the protection policy currently in force for each session matching the request.

dgsdrsNextToken :: Lens' DescribeGameSessionDetailsResponse (Maybe Text)

Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dgsdrsResponseStatus :: Lens' DescribeGameSessionDetailsResponse Int

- | The response status code.