Network.AWS.GameLift.DescribeGameSessionQueues

Description

Retrieves the properties for one or more game session queues. When requesting multiple queues, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a GameSessionQueue object is returned for each requested queue. When specifying a list of queues, objects are returned only for queues that currently exist in the region.

Queue-related operations include:

CreateGameSessionQueue

DescribeGameSessionQueues

UpdateGameSessionQueue

DeleteGameSessionQueue

Creating a Request

describeGameSessionQueues :: DescribeGameSessionQueues

Creates a value of DescribeGameSessionQueues with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dgsqNextToken - Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

dgsqNames - List of queue names to retrieve information for. To request settings for all queues, leave this parameter empty.

dgsqLimit - Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

data DescribeGameSessionQueues

Represents the input for a request action.

See: describeGameSessionQueues smart constructor.

Request Lenses

dgsqNextToken :: Lens' DescribeGameSessionQueues (Maybe Text)

Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

dgsqNames :: Lens' DescribeGameSessionQueues [Text]

List of queue names to retrieve information for. To request settings for all queues, leave this parameter empty.

dgsqLimit :: Lens' DescribeGameSessionQueues (Maybe Natural)

Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Destructuring the Response

describeGameSessionQueuesResponse

Arguments

:: Int
drsResponseStatus

-> DescribeGameSessionQueuesResponse

Creates a value of DescribeGameSessionQueuesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsNextToken - Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

drsGameSessionQueues - Collection of objects that describes the requested game session queues.

drsResponseStatus - -- | The response status code.

data DescribeGameSessionQueuesResponse

Represents the returned data in response to a request action.

See: describeGameSessionQueuesResponse smart constructor.

Response Lenses

drsNextToken :: Lens' DescribeGameSessionQueuesResponse (Maybe Text)

Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

drsGameSessionQueues :: Lens' DescribeGameSessionQueuesResponse [GameSessionQueue]

Collection of objects that describes the requested game session queues.

drsResponseStatus :: Lens' DescribeGameSessionQueuesResponse Int

- | The response status code.