Network.AWS.GameLift.DescribeGameSessions

Description

Retrieves a set of one or more game sessions. Request a specific game session or request all game sessions on a fleet. Alternatively, use SearchGameSessions to request a set of active game sessions that are filtered by certain criteria. To retrieve protection policy settings for game sessions, use DescribeGameSessionDetails .

To get game sessions, specify one of the following: game session ID, fleet ID, or alias ID. You can filter this request by game session status. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, a GameSession object is returned for each game session matching the request.

Available in Amazon GameLift Local.

Game-session-related operations include:

CreateGameSession

DescribeGameSessions

DescribeGameSessionDetails

SearchGameSessions

UpdateGameSession

GetGameSessionLogUrl

Game session placements

StartGameSessionPlacement

DescribeGameSessionPlacement

StopGameSessionPlacement

Creating a Request

describeGameSessions :: DescribeGameSessions

Creates a value of DescribeGameSessions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dgsGameSessionId - Unique identifier for the game session to retrieve. You can use either a GameSessionId or GameSessionArn value.

dgsAliasId - Unique identifier for an alias associated with the fleet to retrieve all game sessions for.

dgsNextToken - Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

dgsStatusFilter - Game session status to filter results on. Possible game session statuses include ACTIVE , TERMINATED , ACTIVATING , and TERMINATING (the last two are transitory).

dgsLimit - Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

dgsFleetId - Unique identifier for a fleet to retrieve all game sessions for.

data DescribeGameSessions

Represents the input for a request action.

See: describeGameSessions smart constructor.

Request Lenses

dgsGameSessionId :: Lens' DescribeGameSessions (Maybe Text)

Unique identifier for the game session to retrieve. You can use either a GameSessionId or GameSessionArn value.

dgsAliasId :: Lens' DescribeGameSessions (Maybe Text)

Unique identifier for an alias associated with the fleet to retrieve all game sessions for.

dgsNextToken :: Lens' DescribeGameSessions (Maybe Text)

Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

dgsStatusFilter :: Lens' DescribeGameSessions (Maybe Text)

Game session status to filter results on. Possible game session statuses include ACTIVE , TERMINATED , ACTIVATING , and TERMINATING (the last two are transitory).

dgsLimit :: Lens' DescribeGameSessions (Maybe Natural)

Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

dgsFleetId :: Lens' DescribeGameSessions (Maybe Text)

Unique identifier for a fleet to retrieve all game sessions for.

Destructuring the Response

describeGameSessionsResponse

Arguments

:: Int
dgsrsResponseStatus

-> DescribeGameSessionsResponse

Creates a value of DescribeGameSessionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dgsrsGameSessions - Collection of objects containing game session properties for each session matching the request.

dgsrsNextToken - Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dgsrsResponseStatus - -- | The response status code.

data DescribeGameSessionsResponse

Represents the returned data in response to a request action.

See: describeGameSessionsResponse smart constructor.

Response Lenses

dgsrsGameSessions :: Lens' DescribeGameSessionsResponse [GameSession]

Collection of objects containing game session properties for each session matching the request.

dgsrsNextToken :: Lens' DescribeGameSessionsResponse (Maybe Text)

Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dgsrsResponseStatus :: Lens' DescribeGameSessionsResponse Int

- | The response status code.