Network.AWS.GameLift.DescribeMatchmakingConfigurations

Description

Retrieves the details of FlexMatch matchmaking configurations. with this operation, you have the following options: (1) retrieve all existing configurations, (2) provide the names of one or more configurations to retrieve, or (3) retrieve all configurations that use a specified rule set name. When requesting multiple items, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a configuration is returned for each requested name. When specifying a list of names, only configurations that currently exist are returned.

Operations related to match configurations and rule sets include:

CreateMatchmakingConfiguration

DescribeMatchmakingConfigurations

UpdateMatchmakingConfiguration

DeleteMatchmakingConfiguration

CreateMatchmakingRuleSet

DescribeMatchmakingRuleSets

ValidateMatchmakingRuleSet

Creating a Request

describeMatchmakingConfigurations :: DescribeMatchmakingConfigurations

Creates a value of DescribeMatchmakingConfigurations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmcRuleSetName - Unique identifier for a matchmaking rule set. Use this parameter to retrieve all matchmaking configurations that use this rule set.

dmcNextToken - Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

dmcNames - Unique identifier for a matchmaking configuration(s) to retrieve. To request all existing configurations, leave this parameter empty.

dmcLimit - Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is limited to 10.

data DescribeMatchmakingConfigurations

Represents the input for a request action.

See: describeMatchmakingConfigurations smart constructor.

Request Lenses

dmcRuleSetName :: Lens' DescribeMatchmakingConfigurations (Maybe Text)

Unique identifier for a matchmaking rule set. Use this parameter to retrieve all matchmaking configurations that use this rule set.

dmcNextToken :: Lens' DescribeMatchmakingConfigurations (Maybe Text)

Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

dmcNames :: Lens' DescribeMatchmakingConfigurations [Text]

Unique identifier for a matchmaking configuration(s) to retrieve. To request all existing configurations, leave this parameter empty.

dmcLimit :: Lens' DescribeMatchmakingConfigurations (Maybe Natural)

Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages. This parameter is limited to 10.

Destructuring the Response

describeMatchmakingConfigurationsResponse

Arguments

:: Int
dmcsrsResponseStatus

-> DescribeMatchmakingConfigurationsResponse

Creates a value of DescribeMatchmakingConfigurationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmcsrsConfigurations - Collection of requested matchmaking configuration objects.

dmcsrsNextToken - Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dmcsrsResponseStatus - -- | The response status code.

data DescribeMatchmakingConfigurationsResponse

Represents the returned data in response to a request action.

See: describeMatchmakingConfigurationsResponse smart constructor.

Response Lenses

dmcsrsConfigurations :: Lens' DescribeMatchmakingConfigurationsResponse [MatchmakingConfiguration]

Collection of requested matchmaking configuration objects.

dmcsrsNextToken :: Lens' DescribeMatchmakingConfigurationsResponse (Maybe Text)

Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

dmcsrsResponseStatus :: Lens' DescribeMatchmakingConfigurationsResponse Int

- | The response status code.