Network.AWS.GameLift.GetInstanceAccess

Description

Requests remote access to a fleet instance. Remote access is useful for debugging, gathering benchmarking data, or watching activity in real time.

Access requires credentials that match the operating system of the instance. For a Windows instance, Amazon GameLift returns a user name and password as strings for use with a Windows Remote Desktop client. For a Linux instance, Amazon GameLift returns a user name and RSA private key, also as strings, for use with an SSH client. The private key must be saved in the proper format to a .pem file before using. If you're making this request using the AWS CLI, saving the secret can be handled as part of the GetInstanceAccess request. (See the example later in this topic). For more information on remote access, see Remotely Accessing an Instance .

To request access to a specific instance, specify the IDs of the instance and the fleet it belongs to. If successful, an InstanceAccess object is returned containing the instance's IP address and a set of credentials.

Creating a Request

getInstanceAccess

Arguments

:: Text
giaFleetId

-> Text
giaInstanceId

-> GetInstanceAccess

Creates a value of GetInstanceAccess with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giaFleetId - Unique identifier for a fleet that contains the instance you want access to. The fleet can be in any of the following statuses: ACTIVATING , ACTIVE , or ERROR . Fleets with an ERROR status may be accessible for a short time before they are deleted.

giaInstanceId - Unique identifier for an instance you want to get access to. You can access an instance in any status.

data GetInstanceAccess

Represents the input for a request action.

See: getInstanceAccess smart constructor.

Request Lenses

giaFleetId :: Lens' GetInstanceAccess Text

Unique identifier for a fleet that contains the instance you want access to. The fleet can be in any of the following statuses: ACTIVATING , ACTIVE , or ERROR . Fleets with an ERROR status may be accessible for a short time before they are deleted.

giaInstanceId :: Lens' GetInstanceAccess Text

Unique identifier for an instance you want to get access to. You can access an instance in any status.

Destructuring the Response

getInstanceAccessResponse

Arguments

:: Int
giarsResponseStatus

-> GetInstanceAccessResponse

Creates a value of GetInstanceAccessResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giarsInstanceAccess - Object that contains connection information for a fleet instance, including IP address and access credentials.

giarsResponseStatus - -- | The response status code.

data GetInstanceAccessResponse

Represents the returned data in response to a request action.

See: getInstanceAccessResponse smart constructor.

Response Lenses

giarsInstanceAccess :: Lens' GetInstanceAccessResponse (Maybe InstanceAccess)

Object that contains connection information for a fleet instance, including IP address and access credentials.

giarsResponseStatus :: Lens' GetInstanceAccessResponse Int

- | The response status code.