Network.AWS.GameLift.ListAliases

Description

Retrieves all aliases for this AWS account. You can filter the result set by alias name and/or routing strategy type. Use the pagination parameters to retrieve results in sequential pages.

Alias-related operations include:

CreateAlias

ListAliases

DescribeAlias

UpdateAlias

DeleteAlias

ResolveAlias

Creating a Request

listAliases :: ListAliases

Creates a value of ListAliases with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

laRoutingStrategyType - Type of routing to filter results on. Use this parameter to retrieve only aliases of a certain type. To retrieve all aliases, leave this parameter empty. Possible routing types include the following: * SIMPLE – The alias resolves to one specific fleet. Use this type when routing to active fleets. * TERMINAL – The alias does not resolve to a fleet but instead can be used to display a message to the user. A terminal alias throws a TerminalRoutingStrategyException with the RoutingStrategy message embedded.

laNextToken - Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

laName - Descriptive label that is associated with an alias. Alias names do not need to be unique.

laLimit - Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

data ListAliases

Represents the input for a request action.

See: listAliases smart constructor.

Request Lenses

laRoutingStrategyType :: Lens' ListAliases (Maybe RoutingStrategyType)

Type of routing to filter results on. Use this parameter to retrieve only aliases of a certain type. To retrieve all aliases, leave this parameter empty. Possible routing types include the following: * SIMPLE – The alias resolves to one specific fleet. Use this type when routing to active fleets. * TERMINAL – The alias does not resolve to a fleet but instead can be used to display a message to the user. A terminal alias throws a TerminalRoutingStrategyException with the RoutingStrategy message embedded.

laNextToken :: Lens' ListAliases (Maybe Text)

Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.

laName :: Lens' ListAliases (Maybe Text)

Descriptive label that is associated with an alias. Alias names do not need to be unique.

laLimit :: Lens' ListAliases (Maybe Natural)

Maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

Destructuring the Response

listAliasesResponse

Arguments

:: Int
larsResponseStatus

-> ListAliasesResponse

Creates a value of ListAliasesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

larsAliases - Collection of alias records that match the list request.

larsNextToken - Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

larsResponseStatus - -- | The response status code.

data ListAliasesResponse

Represents the returned data in response to a request action.

See: listAliasesResponse smart constructor.

Response Lenses

larsAliases :: Lens' ListAliasesResponse [Alias]

Collection of alias records that match the list request.

larsNextToken :: Lens' ListAliasesResponse (Maybe Text)

Token that indicates where to resume retrieving results on the next call to this action. If no token is returned, these results represent the end of the list.

larsResponseStatus :: Lens' ListAliasesResponse Int

- | The response status code.