Network.AWS.GameLift.RequestUploadCredentials

Description

This API call is not currently in use. Retrieves a fresh set of upload credentials and the assigned Amazon S3 storage location for a specific build. Valid credentials are required to upload your game build files to Amazon S3.

Creating a Request

requestUploadCredentials

Arguments

:: Text
rucBuildId

-> RequestUploadCredentials

Creates a value of RequestUploadCredentials with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rucBuildId - Unique identifier for a build to get credentials for.

data RequestUploadCredentials

Represents the input for a request action.

See: requestUploadCredentials smart constructor.

Request Lenses

rucBuildId :: Lens' RequestUploadCredentials Text

Unique identifier for a build to get credentials for.

Destructuring the Response

requestUploadCredentialsResponse

Arguments

:: Int
rucrsResponseStatus

-> RequestUploadCredentialsResponse

Creates a value of RequestUploadCredentialsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rucrsStorageLocation - Amazon S3 path and key, identifying where the game build files are stored.

rucrsUploadCredentials - AWS credentials required when uploading a game build to the storage location. These credentials have a limited lifespan and are valid only for the build they were issued for.

rucrsResponseStatus - -- | The response status code.

data RequestUploadCredentialsResponse

Represents the returned data in response to a request action.

See: requestUploadCredentialsResponse smart constructor.

Response Lenses

rucrsStorageLocation :: Lens' RequestUploadCredentialsResponse (Maybe S3Location)

Amazon S3 path and key, identifying where the game build files are stored.

rucrsUploadCredentials :: Lens' RequestUploadCredentialsResponse (Maybe AWSCredentials)

AWS credentials required when uploading a game build to the storage location. These credentials have a limited lifespan and are valid only for the build they were issued for.

rucrsResponseStatus :: Lens' RequestUploadCredentialsResponse Int

- | The response status code.