Network.AWS.GameLift.ResolveAlias

Description

Retrieves the fleet ID that a specified alias is currently pointing to.

Alias-related operations include:

CreateAlias

ListAliases

DescribeAlias

UpdateAlias

DeleteAlias

ResolveAlias

Creating a Request

resolveAlias

Arguments

:: Text
raAliasId

-> ResolveAlias

Creates a value of ResolveAlias with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

raAliasId - Unique identifier for the alias you want to resolve.

data ResolveAlias

Represents the input for a request action.

See: resolveAlias smart constructor.

Request Lenses

raAliasId :: Lens' ResolveAlias Text

Unique identifier for the alias you want to resolve.

Destructuring the Response

resolveAliasResponse

Arguments

:: Int
rarsResponseStatus

-> ResolveAliasResponse

Creates a value of ResolveAliasResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rarsFleetId - Fleet identifier that is associated with the requested alias.

rarsResponseStatus - -- | The response status code.

data ResolveAliasResponse

Represents the returned data in response to a request action.

See: resolveAliasResponse smart constructor.

Response Lenses

rarsFleetId :: Lens' ResolveAliasResponse (Maybe Text)

Fleet identifier that is associated with the requested alias.

rarsResponseStatus :: Lens' ResolveAliasResponse Int

- | The response status code.