Network.AWS.GameLift.StopGameSessionPlacement

Description

Cancels a game session placement that is in PENDING status. To stop a placement, provide the placement ID values. If successful, the placement is moved to CANCELLED status.

Game-session-related operations include:

CreateGameSession

DescribeGameSessions

DescribeGameSessionDetails

SearchGameSessions

UpdateGameSession

GetGameSessionLogUrl

Game session placements

StartGameSessionPlacement

DescribeGameSessionPlacement

StopGameSessionPlacement

Creating a Request

stopGameSessionPlacement

Arguments

:: Text
sPlacementId

-> StopGameSessionPlacement

Creates a value of StopGameSessionPlacement with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sPlacementId - Unique identifier for a game session placement to cancel.

data StopGameSessionPlacement

Represents the input for a request action.

See: stopGameSessionPlacement smart constructor.

Request Lenses

sPlacementId :: Lens' StopGameSessionPlacement Text

Unique identifier for a game session placement to cancel.

Destructuring the Response

stopGameSessionPlacementResponse

Arguments

:: Int
storsResponseStatus

-> StopGameSessionPlacementResponse

Creates a value of StopGameSessionPlacementResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

storsGameSessionPlacement - Object that describes the canceled game session placement, with CANCELLED status and an end time stamp.

storsResponseStatus - -- | The response status code.

data StopGameSessionPlacementResponse

Represents the returned data in response to a request action.

See: stopGameSessionPlacementResponse smart constructor.

Response Lenses

storsGameSessionPlacement :: Lens' StopGameSessionPlacementResponse (Maybe GameSessionPlacement)

Object that describes the canceled game session placement, with CANCELLED status and an end time stamp.

storsResponseStatus :: Lens' StopGameSessionPlacementResponse Int

- | The response status code.