Network.AWS.GameLift.UpdateAlias

Description

Updates properties for an alias. To update properties, specify the alias ID to be updated and provide the information to be changed. To reassign an alias to another fleet, provide an updated routing strategy. If successful, the updated alias record is returned.

Alias-related operations include:

CreateAlias

ListAliases

DescribeAlias

UpdateAlias

DeleteAlias

ResolveAlias

Creating a Request

updateAlias

Arguments

:: Text
uaAliasId

-> UpdateAlias

Creates a value of UpdateAlias with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uaRoutingStrategy - Object that specifies the fleet and routing type to use for the alias.

uaName - Descriptive label that is associated with an alias. Alias names do not need to be unique.

uaDescription - Human-readable description of an alias.

uaAliasId - Unique identifier for a fleet alias. Specify the alias you want to update.

data UpdateAlias

Represents the input for a request action.

See: updateAlias smart constructor.

Request Lenses

uaRoutingStrategy :: Lens' UpdateAlias (Maybe RoutingStrategy)

Object that specifies the fleet and routing type to use for the alias.

uaName :: Lens' UpdateAlias (Maybe Text)

Descriptive label that is associated with an alias. Alias names do not need to be unique.

uaDescription :: Lens' UpdateAlias (Maybe Text)

Human-readable description of an alias.

uaAliasId :: Lens' UpdateAlias Text

Unique identifier for a fleet alias. Specify the alias you want to update.

Destructuring the Response

updateAliasResponse

Arguments

:: Int
uarsResponseStatus

-> UpdateAliasResponse

Creates a value of UpdateAliasResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uarsAlias - Object that contains the updated alias configuration.

uarsResponseStatus - -- | The response status code.

data UpdateAliasResponse

Represents the returned data in response to a request action.

See: updateAliasResponse smart constructor.

Response Lenses

uarsAlias :: Lens' UpdateAliasResponse (Maybe Alias)

Object that contains the updated alias configuration.

uarsResponseStatus :: Lens' UpdateAliasResponse Int

- | The response status code.