Network.AWS.GameLift.UpdateBuild

Description

Updates metadata in a build record, including the build name and version. To update the metadata, specify the build ID to update and provide the new values. If successful, a build object containing the updated metadata is returned.

Build-related operations include:

CreateBuild

ListBuilds

DescribeBuild

UpdateBuild

DeleteBuild

Creating a Request

updateBuild

Arguments

:: Text
ubBuildId

-> UpdateBuild

Creates a value of UpdateBuild with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ubName - Descriptive label that is associated with a build. Build names do not need to be unique.

ubVersion - Version that is associated with this build. Version strings do not need to be unique.

ubBuildId - Unique identifier for a build to update.

data UpdateBuild

Represents the input for a request action.

See: updateBuild smart constructor.

Request Lenses

ubName :: Lens' UpdateBuild (Maybe Text)

Descriptive label that is associated with a build. Build names do not need to be unique.

ubVersion :: Lens' UpdateBuild (Maybe Text)

Version that is associated with this build. Version strings do not need to be unique.

ubBuildId :: Lens' UpdateBuild Text

Unique identifier for a build to update.

Destructuring the Response

updateBuildResponse

Arguments

:: Int
ubrsResponseStatus

-> UpdateBuildResponse

Creates a value of UpdateBuildResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ubrsBuild - Object that contains the updated build record.

ubrsResponseStatus - -- | The response status code.

data UpdateBuildResponse

Represents the returned data in response to a request action.

See: updateBuildResponse smart constructor.

Response Lenses

ubrsBuild :: Lens' UpdateBuildResponse (Maybe Build)

Object that contains the updated build record.

ubrsResponseStatus :: Lens' UpdateBuildResponse Int

- | The response status code.