Network.AWS.GameLift.UpdateFleetAttributes

Description

Updates fleet properties, including name and description, for a fleet. To update metadata, specify the fleet ID and the property values that you want to change. If successful, the fleet ID for the updated fleet is returned.

Fleet-related operations include:

CreateFleet

ListFleets

Describe fleets:

DescribeFleetAttributes

DescribeFleetPortSettings

DescribeFleetUtilization

DescribeRuntimeConfiguration

DescribeFleetEvents

Update fleets:

UpdateFleetAttributes

UpdateFleetCapacity

UpdateFleetPortSettings

UpdateRuntimeConfiguration

Manage fleet capacity:

DescribeFleetCapacity

UpdateFleetCapacity

PutScalingPolicy (automatic scaling)

DescribeScalingPolicies (automatic scaling)

DeleteScalingPolicy (automatic scaling)

DescribeEC2InstanceLimits

DeleteFleet

Creating a Request

updateFleetAttributes

Arguments

:: Text
ufaFleetId

-> UpdateFleetAttributes

Creates a value of UpdateFleetAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ufaNewGameSessionProtectionPolicy - Game session protection policy to apply to all new instances created in this fleet. Instances that already exist are not affected. You can set protection for individual instances using UpdateGameSession . * NoProtection – The game session can be terminated during a scale-down event. * FullProtection – If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

ufaName - Descriptive label that is associated with a fleet. Fleet names do not need to be unique.

ufaMetricGroups - Names of metric groups to include this fleet in. Amazon CloudWatch uses a fleet metric group is to aggregate metrics from multiple fleets. Use an existing metric group name to add this fleet to the group. Or use a new name to create a new metric group. A fleet can only be included in one metric group at a time.

ufaDescription - Human-readable description of a fleet.

ufaResourceCreationLimitPolicy - Policy that limits the number of game sessions an individual player can create over a span of time.

ufaFleetId - Unique identifier for a fleet to update attribute metadata for.

data UpdateFleetAttributes

Represents the input for a request action.

See: updateFleetAttributes smart constructor.

Request Lenses

ufaNewGameSessionProtectionPolicy :: Lens' UpdateFleetAttributes (Maybe ProtectionPolicy)

Game session protection policy to apply to all new instances created in this fleet. Instances that already exist are not affected. You can set protection for individual instances using UpdateGameSession . * NoProtection – The game session can be terminated during a scale-down event. * FullProtection – If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

ufaName :: Lens' UpdateFleetAttributes (Maybe Text)

Descriptive label that is associated with a fleet. Fleet names do not need to be unique.

ufaMetricGroups :: Lens' UpdateFleetAttributes [Text]

Names of metric groups to include this fleet in. Amazon CloudWatch uses a fleet metric group is to aggregate metrics from multiple fleets. Use an existing metric group name to add this fleet to the group. Or use a new name to create a new metric group. A fleet can only be included in one metric group at a time.

ufaDescription :: Lens' UpdateFleetAttributes (Maybe Text)

Human-readable description of a fleet.

ufaResourceCreationLimitPolicy :: Lens' UpdateFleetAttributes (Maybe ResourceCreationLimitPolicy)

Policy that limits the number of game sessions an individual player can create over a span of time.

ufaFleetId :: Lens' UpdateFleetAttributes Text

Unique identifier for a fleet to update attribute metadata for.

Destructuring the Response

updateFleetAttributesResponse

Arguments

:: Int
ufarsResponseStatus

-> UpdateFleetAttributesResponse

Creates a value of UpdateFleetAttributesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ufarsFleetId - Unique identifier for a fleet that was updated.

ufarsResponseStatus - -- | The response status code.

data UpdateFleetAttributesResponse

Represents the returned data in response to a request action.

See: updateFleetAttributesResponse smart constructor.

Response Lenses

ufarsFleetId :: Lens' UpdateFleetAttributesResponse (Maybe Text)

Unique identifier for a fleet that was updated.

ufarsResponseStatus :: Lens' UpdateFleetAttributesResponse Int

- | The response status code.