Network.AWS.GameLift.UpdateFleetPortSettings

Description

Updates port settings for a fleet. To update settings, specify the fleet ID to be updated and list the permissions you want to update. List the permissions you want to add in InboundPermissionAuthorizations , and permissions you want to remove in InboundPermissionRevocations . Permissions to be removed must match existing fleet permissions. If successful, the fleet ID for the updated fleet is returned.

Fleet-related operations include:

CreateFleet

ListFleets

Describe fleets:

DescribeFleetAttributes

DescribeFleetPortSettings

DescribeFleetUtilization

DescribeRuntimeConfiguration

DescribeFleetEvents

Update fleets:

UpdateFleetAttributes

UpdateFleetCapacity

UpdateFleetPortSettings

UpdateRuntimeConfiguration

Manage fleet capacity:

DescribeFleetCapacity

UpdateFleetCapacity

PutScalingPolicy (automatic scaling)

DescribeScalingPolicies (automatic scaling)

DeleteScalingPolicy (automatic scaling)

DescribeEC2InstanceLimits

DeleteFleet

Creating a Request

updateFleetPortSettings

Arguments

:: Text
ufpsFleetId

-> UpdateFleetPortSettings

Creates a value of UpdateFleetPortSettings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ufpsInboundPermissionRevocations - Collection of port settings to be removed from the fleet record.

ufpsInboundPermissionAuthorizations - Collection of port settings to be added to the fleet record.

ufpsFleetId - Unique identifier for a fleet to update port settings for.

data UpdateFleetPortSettings

Represents the input for a request action.

See: updateFleetPortSettings smart constructor.

Request Lenses

ufpsInboundPermissionRevocations :: Lens' UpdateFleetPortSettings [IPPermission]

Collection of port settings to be removed from the fleet record.

ufpsInboundPermissionAuthorizations :: Lens' UpdateFleetPortSettings [IPPermission]

Collection of port settings to be added to the fleet record.

ufpsFleetId :: Lens' UpdateFleetPortSettings Text

Unique identifier for a fleet to update port settings for.

Destructuring the Response

updateFleetPortSettingsResponse

Arguments

:: Int
ufpsrsResponseStatus

-> UpdateFleetPortSettingsResponse

Creates a value of UpdateFleetPortSettingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ufpsrsFleetId - Unique identifier for a fleet that was updated.

ufpsrsResponseStatus - -- | The response status code.

data UpdateFleetPortSettingsResponse

Represents the returned data in response to a request action.

See: updateFleetPortSettingsResponse smart constructor.

Response Lenses

ufpsrsFleetId :: Lens' UpdateFleetPortSettingsResponse (Maybe Text)

Unique identifier for a fleet that was updated.

ufpsrsResponseStatus :: Lens' UpdateFleetPortSettingsResponse Int

- | The response status code.