Network.AWS.GameLift.UpdateGameSession

Description

Updates game session properties. This includes the session name, maximum player count, protection policy, which controls whether or not an active game session can be terminated during a scale-down event, and the player session creation policy, which controls whether or not new players can join the session. To update a game session, specify the game session ID and the values you want to change. If successful, an updated GameSession object is returned.

Game-session-related operations include:

CreateGameSession

DescribeGameSessions

DescribeGameSessionDetails

SearchGameSessions

UpdateGameSession

GetGameSessionLogUrl

Game session placements

StartGameSessionPlacement

DescribeGameSessionPlacement

StopGameSessionPlacement

Creating a Request

updateGameSession

Arguments

:: Text
ugsGameSessionId

-> UpdateGameSession

Creates a value of UpdateGameSession with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ugsMaximumPlayerSessionCount - Maximum number of players that can be connected simultaneously to the game session.

ugsPlayerSessionCreationPolicy - Policy determining whether or not the game session accepts new players.

ugsName - Descriptive label that is associated with a game session. Session names do not need to be unique.

ugsProtectionPolicy - Game session protection policy to apply to this game session only. * NoProtection – The game session can be terminated during a scale-down event. * FullProtection – If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

ugsGameSessionId - Unique identifier for the game session to update.

data UpdateGameSession

Represents the input for a request action.

See: updateGameSession smart constructor.

Request Lenses

ugsMaximumPlayerSessionCount :: Lens' UpdateGameSession (Maybe Natural)

Maximum number of players that can be connected simultaneously to the game session.

ugsPlayerSessionCreationPolicy :: Lens' UpdateGameSession (Maybe PlayerSessionCreationPolicy)

Policy determining whether or not the game session accepts new players.

ugsName :: Lens' UpdateGameSession (Maybe Text)

Descriptive label that is associated with a game session. Session names do not need to be unique.

ugsProtectionPolicy :: Lens' UpdateGameSession (Maybe ProtectionPolicy)

Game session protection policy to apply to this game session only. * NoProtection – The game session can be terminated during a scale-down event. * FullProtection – If the game session is in an ACTIVE status, it cannot be terminated during a scale-down event.

ugsGameSessionId :: Lens' UpdateGameSession Text

Unique identifier for the game session to update.

Destructuring the Response

updateGameSessionResponse

Arguments

:: Int
ugsrsResponseStatus

-> UpdateGameSessionResponse

Creates a value of UpdateGameSessionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ugsrsGameSession - Object that contains the updated game session metadata.

ugsrsResponseStatus - -- | The response status code.

data UpdateGameSessionResponse

Represents the returned data in response to a request action.

See: updateGameSessionResponse smart constructor.

Response Lenses

ugsrsGameSession :: Lens' UpdateGameSessionResponse (Maybe GameSession)

Object that contains the updated game session metadata.

ugsrsResponseStatus :: Lens' UpdateGameSessionResponse Int

- | The response status code.