Network.AWS.GameLift.UpdateMatchmakingConfiguration

Description

Updates settings for a FlexMatch matchmaking configuration. To update settings, specify the configuration name to be updated and provide the new settings.

Operations related to match configurations and rule sets include:

CreateMatchmakingConfiguration

DescribeMatchmakingConfigurations

UpdateMatchmakingConfiguration

DeleteMatchmakingConfiguration

CreateMatchmakingRuleSet

DescribeMatchmakingRuleSets

ValidateMatchmakingRuleSet

Creating a Request

updateMatchmakingConfiguration

Arguments

:: Text
umcName

-> UpdateMatchmakingConfiguration

Creates a value of UpdateMatchmakingConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umcGameProperties - Set of developer-defined properties for a game session, formatted as a set of type:value pairs. These properties are included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ). This information is added to the new GameSession object that is created for a successful match.

umcRuleSetName - Unique identifier for a matchmaking rule set to use with this configuration. A matchmaking configuration can only use rule sets that are defined in the same region.

umcAcceptanceTimeoutSeconds - Length of time (in seconds) to wait for players to accept a proposed match. If any player rejects the match or fails to accept before the timeout, the ticket continues to look for an acceptable match.

umcRequestTimeoutSeconds - Maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that time out can be resubmitted as needed.

umcNotificationTarget - SNS topic ARN that is set up to receive matchmaking notifications. See Setting up Notifications for Matchmaking for more information.

umcGameSessionQueueARNs - Amazon Resource Name (ARN ) that is assigned to a game session queue and uniquely identifies it. Format is arn:aws:gamelift:region::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912 . These queues are used when placing game sessions for matches that are created with this matchmaking configuration. Queues can be located in any region.

umcCustomEventData - Information to attached to all events related to the matchmaking configuration.

umcAcceptanceRequired - Flag that determines whether or not a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE.

umcGameSessionData - Set of developer-defined game session properties, formatted as a single string value. This data is included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ). This information is added to the new GameSession object that is created for a successful match.

umcDescription - Descriptive label that is associated with matchmaking configuration.

umcAdditionalPlayerCount - Number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match.

umcName - Unique identifier for a matchmaking configuration to update.

data UpdateMatchmakingConfiguration

Represents the input for a request action.

See: updateMatchmakingConfiguration smart constructor.

Request Lenses

umcGameProperties :: Lens' UpdateMatchmakingConfiguration [GameProperty]

Set of developer-defined properties for a game session, formatted as a set of type:value pairs. These properties are included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ). This information is added to the new GameSession object that is created for a successful match.

umcRuleSetName :: Lens' UpdateMatchmakingConfiguration (Maybe Text)

Unique identifier for a matchmaking rule set to use with this configuration. A matchmaking configuration can only use rule sets that are defined in the same region.

umcAcceptanceTimeoutSeconds :: Lens' UpdateMatchmakingConfiguration (Maybe Natural)

Length of time (in seconds) to wait for players to accept a proposed match. If any player rejects the match or fails to accept before the timeout, the ticket continues to look for an acceptable match.

umcRequestTimeoutSeconds :: Lens' UpdateMatchmakingConfiguration (Maybe Natural)

Maximum duration, in seconds, that a matchmaking ticket can remain in process before timing out. Requests that time out can be resubmitted as needed.

umcNotificationTarget :: Lens' UpdateMatchmakingConfiguration (Maybe Text)

SNS topic ARN that is set up to receive matchmaking notifications. See Setting up Notifications for Matchmaking for more information.

umcGameSessionQueueARNs :: Lens' UpdateMatchmakingConfiguration [Text]

Amazon Resource Name (ARN ) that is assigned to a game session queue and uniquely identifies it. Format is arn:aws:gamelift:region::fleet/fleet-a1234567-b8c9-0d1e-2fa3-b45c6d7e8912 . These queues are used when placing game sessions for matches that are created with this matchmaking configuration. Queues can be located in any region.

umcCustomEventData :: Lens' UpdateMatchmakingConfiguration (Maybe Text)

Information to attached to all events related to the matchmaking configuration.

umcAcceptanceRequired :: Lens' UpdateMatchmakingConfiguration (Maybe Bool)

Flag that determines whether or not a match that was created with this configuration must be accepted by the matched players. To require acceptance, set to TRUE.

umcGameSessionData :: Lens' UpdateMatchmakingConfiguration (Maybe Text)

Set of developer-defined game session properties, formatted as a single string value. This data is included in the GameSession object, which is passed to the game server with a request to start a new game session (see Start a Game Session ). This information is added to the new GameSession object that is created for a successful match.

umcDescription :: Lens' UpdateMatchmakingConfiguration (Maybe Text)

Descriptive label that is associated with matchmaking configuration.

umcAdditionalPlayerCount :: Lens' UpdateMatchmakingConfiguration (Maybe Natural)

Number of player slots in a match to keep open for future players. For example, if the configuration's rule set specifies a match for a single 12-person team, and the additional player count is set to 2, only 10 players are selected for the match.

umcName :: Lens' UpdateMatchmakingConfiguration Text

Unique identifier for a matchmaking configuration to update.

Destructuring the Response

updateMatchmakingConfigurationResponse

Arguments

:: Int
umcrsResponseStatus

-> UpdateMatchmakingConfigurationResponse

Creates a value of UpdateMatchmakingConfigurationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umcrsConfiguration - Object that describes the updated matchmaking configuration.

umcrsResponseStatus - -- | The response status code.

data UpdateMatchmakingConfigurationResponse

Represents the returned data in response to a request action.

See: updateMatchmakingConfigurationResponse smart constructor.

Response Lenses

umcrsConfiguration :: Lens' UpdateMatchmakingConfigurationResponse (Maybe MatchmakingConfiguration)

Object that describes the updated matchmaking configuration.

umcrsResponseStatus :: Lens' UpdateMatchmakingConfigurationResponse Int

- | The response status code.