Network.AWS.GameLift.UpdateRuntimeConfiguration

Description

Updates the current run-time configuration for the specified fleet, which tells Amazon GameLift how to launch server processes on instances in the fleet. You can update a fleet's run-time configuration at any time after the fleet is created; it does not need to be in an ACTIVE status.

To update run-time configuration, specify the fleet ID and provide a RuntimeConfiguration object with the updated collection of server process configurations.

Each instance in a Amazon GameLift fleet checks regularly for an updated run-time configuration and changes how it launches server processes to comply with the latest version. Existing server processes are not affected by the update; they continue to run until they end, while Amazon GameLift simply adds new server processes to fit the current run-time configuration. As a result, the run-time configuration changes are applied gradually as existing processes shut down and new processes are launched in Amazon GameLift's normal process recycling activity.

Fleet-related operations include:

CreateFleet

ListFleets

Describe fleets:

DescribeFleetAttributes

DescribeFleetPortSettings

DescribeFleetUtilization

DescribeRuntimeConfiguration

DescribeFleetEvents

Update fleets:

UpdateFleetAttributes

UpdateFleetCapacity

UpdateFleetPortSettings

UpdateRuntimeConfiguration

Manage fleet capacity:

DescribeFleetCapacity

UpdateFleetCapacity

PutScalingPolicy (automatic scaling)

DescribeScalingPolicies (automatic scaling)

DeleteScalingPolicy (automatic scaling)

DescribeEC2InstanceLimits

DeleteFleet

Creating a Request

updateRuntimeConfiguration

Arguments

:: Text
urcFleetId

-> RuntimeConfiguration
urcRuntimeConfiguration

-> UpdateRuntimeConfiguration

Creates a value of UpdateRuntimeConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urcFleetId - Unique identifier for a fleet to update run-time configuration for.

urcRuntimeConfiguration - Instructions for launching server processes on each instance in the fleet. The run-time configuration for a fleet has a collection of server process configurations, one for each type of server process to run on an instance. A server process configuration specifies the location of the server executable, launch parameters, and the number of concurrent processes with that configuration to maintain on each instance.

data UpdateRuntimeConfiguration

Represents the input for a request action.

See: updateRuntimeConfiguration smart constructor.

Request Lenses

urcFleetId :: Lens' UpdateRuntimeConfiguration Text

Unique identifier for a fleet to update run-time configuration for.

urcRuntimeConfiguration :: Lens' UpdateRuntimeConfiguration RuntimeConfiguration

Instructions for launching server processes on each instance in the fleet. The run-time configuration for a fleet has a collection of server process configurations, one for each type of server process to run on an instance. A server process configuration specifies the location of the server executable, launch parameters, and the number of concurrent processes with that configuration to maintain on each instance.

Destructuring the Response

updateRuntimeConfigurationResponse

Arguments

:: Int
urcrsResponseStatus

-> UpdateRuntimeConfigurationResponse

Creates a value of UpdateRuntimeConfigurationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urcrsRuntimeConfiguration - The run-time configuration currently in force. If the update was successful, this object matches the one in the request.

urcrsResponseStatus - -- | The response status code.

data UpdateRuntimeConfigurationResponse

Represents the returned data in response to a request action.

See: updateRuntimeConfigurationResponse smart constructor.

Response Lenses

urcrsRuntimeConfiguration :: Lens' UpdateRuntimeConfigurationResponse (Maybe RuntimeConfiguration)

The run-time configuration currently in force. If the update was successful, this object matches the one in the request.

urcrsResponseStatus :: Lens' UpdateRuntimeConfigurationResponse Int

- | The response status code.