Network.AWS.GameLift.ValidateMatchmakingRuleSet

Description

Validates the syntax of a matchmaking rule or rule set. This operation checks that the rule set uses syntactically correct JSON and that it conforms to allowed property expressions. To validate syntax, provide a rule set string.

Operations related to match configurations and rule sets include:

CreateMatchmakingConfiguration

DescribeMatchmakingConfigurations

UpdateMatchmakingConfiguration

DeleteMatchmakingConfiguration

CreateMatchmakingRuleSet

DescribeMatchmakingRuleSets

ValidateMatchmakingRuleSet

Creating a Request

validateMatchmakingRuleSet

Arguments

:: Text
vmrsRuleSetBody

-> ValidateMatchmakingRuleSet

Creates a value of ValidateMatchmakingRuleSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vmrsRuleSetBody - Collection of matchmaking rules to validate, formatted as a JSON string.

data ValidateMatchmakingRuleSet

Represents the input for a request action.

See: validateMatchmakingRuleSet smart constructor.

Request Lenses

vmrsRuleSetBody :: Lens' ValidateMatchmakingRuleSet Text

Collection of matchmaking rules to validate, formatted as a JSON string.

Destructuring the Response

validateMatchmakingRuleSetResponse

Arguments

:: Int
vmrsrsResponseStatus

-> ValidateMatchmakingRuleSetResponse

Creates a value of ValidateMatchmakingRuleSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vmrsrsValid - Response indicating whether or not the rule set is valid.

vmrsrsResponseStatus - -- | The response status code.

data ValidateMatchmakingRuleSetResponse

Represents the returned data in response to a request action.

See: validateMatchmakingRuleSetResponse smart constructor.

Response Lenses

vmrsrsValid :: Lens' ValidateMatchmakingRuleSetResponse (Maybe Bool)

Response indicating whether or not the rule set is valid.

vmrsrsResponseStatus :: Lens' ValidateMatchmakingRuleSetResponse Int

- | The response status code.